Fixed loading issue
