+++ src/app/(app)/attendance/page.tsx
@@ -5,8 +5,7 @@ import { PageLayout } from "@/app/layout/PageLayout";
 import { AttendanceHeader } from "./components/AttendanceHeader";
 import { Routes } from "@/lib/routes";
 import { getLoggedInUser } from "@/utils/getLoggedInUser";
-import { AttendanceFilter } from "@/common/Filter/AttendanceFilter";
-
+import { MemberFilter } from "@/common/Filter/MemberFilter";
 const breadcrumbs = [
   { href: Routes.Attendance, label: "Attendance" },
   { href: Routes.AddAttendance, label: "Review attendance" },
@@ -20,7 +19,7 @@ const Page = async ({ searchParams }: { searchParams: { id: string } }) => {
   return (
     <>
       <AttendanceHeader breadcrumbs={breadcrumbs} />
-      <AttendanceFilter loginUser={loginUser} />
+      <MemberFilter loginUser={loginUser}  route = {Routes.Attendance}/>
       <PageLayout>
         <AttendanceList attendance={attendance} loginUser={loginUser} />
       </PageLayout>
